Create separate query module for graphql queries
Some graphql queries are duplicated across modules, moving them to a single
location will allow us to reuse the code.
