+++ metrics/github/query.py
@@ -1,0 +1,18 @@
+def query_repos(client):
+    query = """
+    query repos($cursor: String, $org: String!) {
+      organization(login: $org) {
+        repositories(first: 100, after: $cursor) {
+          nodes {
+            name
+            archivedAt
+          }
+          pageInfo {
+              endCursor
+              hasNextPage
+          }
+        }
+      }
+    }
+    """
+    return client.get_query(query, path=["organization", "repositories"])
